+++ rustsol/src/types/primitive.rs
@@ -8,23 +8,23 @@ use crate::utils::vec_u256_to_vec_bytes;
 
 #[derive(Derivative, Default)]
 #[derivative(Debug)]
-pub struct Primitive<const SIZE: u64> {
+pub struct Primitive<const SIZE: usize> {
     __slot: U256,
-    __offset: u8,
+    __offset: usize,
     #[derivative(Debug = "ignore")]
     __slot_getter: Option<Arc<dyn SlotsGetter>>,
 }
 
-impl<const SIZE: u64> Primitive<SIZE> {
+impl<const SIZE: usize> Primitive<SIZE> {
     pub fn slot(&self) -> U256 {
         self.__slot
     }
 
-    pub fn offset(&self) -> u8 {
+    pub fn offset(&self) -> usize {
         self.__offset
     }
 
-    pub fn position(&self) -> (U256, u8, u64) {
+    pub fn position(&self) -> (U256, usize, usize) {
         (self.__slot, self.__offset, SIZE)
     }
 
@@ -39,23 +39,23 @@ impl<const SIZE: u64> Primitive<SIZE> {
     }
 }
 
-impl<const SIZE: u64> Position for Primitive<SIZE> {
-    fn from_position(slot: U256, offset: u8) -> Self {
+impl<const SIZE: usize> Position for Primitive<SIZE> {
+    fn from_position(slot: U256, offset: usize) -> Self {
         Primitive { __slot: slot, __offset: offset, __slot_getter: None }  // Use the conversion from U256 to u64
     }
 
-    fn size() -> u64 {
+    fn size() -> usize {
         SIZE
     }
 }
 
-impl<const SIZE: u64> SlotsGetterSetter for Primitive<SIZE> {
+impl<const SIZE: usize> SlotsGetterSetter for Primitive<SIZE> {
     fn set_slots_getter(&mut self, getter: Arc<dyn SlotsGetter>) {
         self.__slot_getter = Some(getter);
     }
 }
 
-impl<const SIZE: u64> Value for Primitive<SIZE> {
+impl<const SIZE: usize> Value for Primitive<SIZE> {
     // TODO: Change to concrete values type like u64 and bool.
     type ValueType = U256;
